+++ src/components/Results/Results.tsx
@@ -5,8 +5,7 @@ import { Outlet } from 'react-router-dom';
 import { selectLimit } from '../../features/limitSlice';
 import { selectPage } from '../../features/pageSlice';
 import { selectSearch } from '../../features/searchSlice';
-import { useAppSelector } from '../../hooks';
-import { useActions } from '../../hooks/useActions';
+import { useActions, useAppSelector } from '../../hooks';
 import { Data, Person } from '../../interfaces/SWApi';
 import { useGetPeopleQuery } from '../../services/api';
 import Pagination from '../Pagination/Pagination';
