+++ src/hooks/index.ts
@@ -3,3 +3,4 @@ import type { AppDispatch, RootState } from '../store/store';
 
 export const useAppDispatch: () => AppDispatch = useDispatch;
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
+export { default as useActions } from './useActions';
